+++ src/main/java/racinggame/RacingCars.java
@@ -1,5 +1,8 @@
 package racinggame;
 
+import racinggame.utils.MoveValueGenerator;
+import racinggame.view.ResultView;
+
 import java.util.List;
 import java.util.stream.Collectors;
