+++ src/main/java/racinggame/utils/MoveValueGenerator.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.utils;
 
 import java.util.Random;
