+++ src/main/java/racinggame/utils/MoveValueValidator.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.utils;
 
 public class MoveValueValidator {
     public static final int ADVANCED_CONDITION = 4;
