+++ src/test/java/racinggame/RacingCarsTest.java
@@ -3,6 +3,7 @@ package racinggame;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racinggame.view.ResultView;
 
 import java.util.ArrayList;
 import java.util.List;
