+++ src/test/java/racinggame/utils/CarFactoryTest.java
@@ -1,8 +1,9 @@
-package racinggame;
+package racinggame.utils;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racinggame.RacingCars;
 
 import static org.assertj.core.api.Assertions.assertThat;
